+++ app/models/static_member.rb
@@ -1,0 +1,5 @@
+# Represents members of statics
+class StaticMember < ApplicationRecord
+  belongs_to :static
+  belongs_to :character
+end