+++ spec/factories/static_members.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :static_member do
+    association :static
+    association :character
+  end
+end