+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - static show page
 - static management page
+- StaticMember model
 
 ## [0.6.0] - 2019-11-01
 ### Added